update server
added previous route to handle redirect problem with login so that it can redirect the page back to the previous one.
fix some issue with homepage route
change the behavior of login to redirect to previous page
create :imageID route which read individual image using image ID as the route. It will display comment, likes.
add like route to handle like button click which log it into the database.
add comment route which save the comment to database.